Changed object types to be strings instead of enums
Map tiles are now represented by strings instead of int-based
enums.  The database is now slightly more human-readable, and
the types are a bit more data-driven.  The list of name-prefab
pairings is now set in the editor through the inspector, and the
object spawning routines, (and the level editor GUI) are
created automatically.

Change-Id: Ice893c592345f371f5b2d9d277ce8ef3b84e9e4b