Changed how ball spawning works.
Previously, balls were first-class map objects, and their position
was saved when the level was saved.  Now, they are completely
absent during the level editor, and instead the user places a "start
location", which is responsible for spawning the ball whenever the
game is being played, and removing it when the game ends.

Change-Id: Id82c1d75ffe106c719c61db6e4ecc9eb200bbe84