Reworked how some of the network connecting works.
Now, instead of trying to keep the table of all maps
in sync with the server, the game will just focus on the
current map being looked at.

Changed the WaitingForDBLoad state to be a lot more general
as a result.  (It used to only load levelmaps, but now it
can be used as a transition state for fetching any struct
on the rmeote server.)

Change-Id: I55014395816b84a16d50be9e4f78bab3a822986b