Added a Title Screen, Main Menu
Added a title screen, with a placeholder menu.  At the moment, the
only menu option that functions is "editor", which simply launches the
editor, but this gives me something to hang functionality on as I build
in the other features.

Also reworked how user loading happens.  Before it was grafted on
ad-hoc to whatever state happened to be happening first, but now it's
been broken out into its own state, making it a lot easier to handle
for control flow.

Also fixed a minor bug in the state manager where states were not
having their values initialized correctly when swapped to.

Change-Id: I0ffe8f9d1bc81d451e28c14e30b8407304ff8d6b