Baked in some basic maps as data
Added some basic maps to the build as json resources, playable
even offline without a database connection.

Also added a menu for selecting them and playing them.

Also added a state (States/ExportMap.cs) to make it easier for me
to generate offline map data and add it to the game.  This class is
a temporary helper class, and is not intended as production code.
It will be removed before launch.  (/b/33822106)

Change-Id: I77efabb4258fea98c82226c59111fc49b82aec8f