Added UI for saving maps.
Adds a screen and UI for selecting a name and saving new maps.
Allows for both the creation of new maps, or saving over
existing ones.

Note:  Because the database location for saved maps has changed
slightly, this CL breaks the ability to load saved maps from
within the app.  This will be fixed in the next checkin, as
part of the UI for selecting a map to load.

Change-Id: I2afeb2ccfa6d37e13142d57d8d8b4b4c9896ce1a