Centralized timescale manipulation
The game sets the timescale to 0 when in menus, to effectively
pause the physics simulation.  (So things don't move while you're
editing them, for example.)

Originally this was done by having every menu ensure the timescale
was 0 at all times.

This is cumbersome, as there are a lot of menus.

It is now handled entirely by the gameplay state, which is
responsible for setting the timescale to 1.0 while it is active,
and 0.0 otherwise.

All other places where timescale was set have been removed.

Change-Id: I342f8efe592317cbe0094a842b47e0f417d144c7