Separated desktop vs. mobile GUI skins
Before, there was an attempt to pick one size for GUI elements
that would work on both desktop vs. mobile resolutions.  This
resulted in sizes that were a bit too big for desktop, and a bit
too small for mobile.

This change splits those out into two separate (independently
editable) skins, and uses the one appropriate to the environment
the game is currently running under.

Change-Id: I9abf4c2eaa1c7c1560f07e6604fc0e578a3fff32