Allowed way to hide prefabs from the tool menu
At the moment, all prefabs registered in the prefablist end
up on the editor menu.  Some of them should not show up on the
menu directly though.  (Example:  The player's hamsterball, which
is a prefab, but is not placed directly in the editor - it is just
spawned when the level starts, from player spawn points.)

This change adds a boolean to the prefab structure, that allows
prefabs to be hidden from the editor list.  Also, minor changes
to how some of the UI lists are pre-populated, since there is no
longer a 1:1 correspondence between the list of prefabs and the
list generated for the menu.

Change-Id: Ie0864352f5aabac8688a36e1c601cf035f52dfb9