Lowering the game world to be on the Z plane.
Elements are added at their centers, which meant everything placed
on the Z plane had the interactible parts above the plane, which
makes editing confusing, as it was based on the plane below what was
visible. By lowering the objects so that the "floor" is now on the
Z plane, it makes editing feel better.

Change-Id: Ib815d1d0080e0bc9bf4cdf1021d27737767a0f74