Added Google VR SDK, changed Unity version
Added the google VR SDK 1.2 to the project, becase we will
need to depend on several of its prefabs.

Source:  https://github.com/googlevr/gvr-unity-sdk

Also updated the project files to unity 5.6 (beta) because the
VR SDK installs extra files if it is running in compatability mode.
(These files are installed automatically if the project is opened
in an older version of Unity, so they don't need to be checked in.)

Change-Id: I04f68d1ea8c29d5ff89e18513dd8405938eb7330