Changing unique tiles to use max count instead.
Adding the new player start tile, which takes the entire tile space, and
thus did not work with the previous unique tile system. Instead there is
now a max count for a tile.
Also adds a script to hide elements during gameplay, used by the giant
arrow on the new player start tile.

Tested on Windows.

Change-Id: I413b4e02801c7a6baf3e577ccecb5faaeebb70e9