First pass at Daydream gameplay.
Since Daydream UI is not in place yet, this checkin also
includes an alternate level path, which only triggers when
VR is enabled, and when running on Android or in the Unity
editor.  This alternate path skips the main menu, and simply
dumps the user into a level.  (And restarts it whenever they
complete it.)

Also started the Daydream functionality as a separate scene,
and moved most things from the main scene into prefabs, so
that they'll be easier to keep in sync.

Change-Id: I93cf667c34019a848ccba997a6458a5da9812c4e