Merged VR and regular scenes.
Combined the scene for VR with the scene for regular mode,
into one large scene.  Added more robust branching logic to
handle this, rather than relying on build settings.

As part of this, also removed the original MainScene, and
replaced it with the (heavily modified) MainScene_DayDream
scene.

Also, fixed a bug in which daydream mode would hang if launched
on devices that hadn't yet signed in in some way.

Change-Id: Id87ee5e13906b6800fd6f7905fdc79833f5c2942