Remove the GVR iOS library, stub out calls.
The GVR iOS library is fully linked, which causes conflicts with the
shared dependencies of it and Firebase, which uses Pods to manage it's
dependencies. To work around this, we remove the GVR iOS library from
being included, as we were not using it, and stub out the calls to the
library.

Tested on iOS.

Change-Id: Iba7580f914c92454d77a88bc87a639956657aaef