Added new unified UI mode to handle VR/non-VR.
This change adds a new system for handling UI across
both VR and non-VR modes.  It no longer uses Unity's
immediate-mode-GUI system, and instead uses Unity's
standard button widgets, projected into world-space.

In this change, the main menu is the only screen that
has been converted, but it now is operable with both
the mouse/finger touches, as well as the daydream
pointer.

Change-Id: I4ca4ae21e93ba09efd37c603e397fb9ae007d849