Converted the basic dialog to the new UI
Basic Dialog is the screen that displays simple messages
to the user.  (Mostly used for reporting errors or events.)
It now works in VR mode, and uses the new UI instead of the old
placeholder UI.

Also changed how UI prefabs are set up - they now include the
canvas, rather than spawning on an existing shared canvas.  This
gives them greater control over the size the panel they're on.

Change-Id: I692347619c0b28122728f868fd63f3d004c42892