Cleaned up camera code.
Now, rather than having the camera try to deduce
how it should behave based on the current state of the
game, it now has an explicit CameraMode enum, which
defines its current behavior.

All states are assumed to be in the Menu camera mode -
any states that change it (gameplay, editor, etc) are
expected to set it back before exiting.

Change-Id: Idedf104d41b90fe1e875defd7644b0368d6525a6