Added new UI for level selection screen.
Enables the level selection screen to work with the new
UI.  (So it now works in both VR and Mobile modes, and
looks more like the intended style.)

Note that this only affects the level selection screen
for builtin levels.  Other level selections (editor loads,
bonus levels, etc) still need to be converted.

Change-Id: I9cbb32a01dba865d609e76aecdf9fe2ebce577be