UI Fixes - Menu Color, and UI size in mobile mode.
When running in mobile (non-VR) mode, the UI was too small.
This change makes the UI spawn closer to the camera, so it
appears larger.  (The UI is rendered in world-space as opposed
to an orthographc camera, to make it easier to combine the logic
between the VR and mobile UIs.)

Also fixed an incorrectly colored menu item.

Change-Id: I9541d19ded93871cabbe8da74554f88600c0e19d