Added VR controller help screen.
When you start a game in VR mode now, it shows a card
explaining how to hold the controller.  This lasts for
a few seconds, but goes away faster if the game thinks
you are holding controller horizontally already.  (i. e. the
accelerometer is almost entirely in the Y axis.)

Also includes minor cleanup to the VR camera code.

Change-Id: Icdec8a899b3c8b0678f4afeca985b1bc5bc11c24