Update the Manage Account UI to the new style.
Updating the Account UI to use the prefab style. Fixes issues in
AddEmail not working. Shrinks the default font size of the BasicDialog
box to better match the types of messages it is used to show.
Hides the Account button on desktop, where it previously caused a bad
state to be entered.

Tested on Windows and Android.

Change-Id: I5f0e469ced758b1eabca4a34d173e057e8900cbe