Minor tweaks to the GUIButton class
Class now has a slightly smoother animation flow,
and no longer has a hard-coded multiplier in the expansion.

From the user's point of view, hover animations should be more
consistent, (always expanding evenly, instead of the behavior
depending on the current system time) and the buttons should
be less huge on press.

Change-Id: I72e80aee5d6f16ea2ef1f8893dc8db25a708efc9