Removed unneeded files.
This was the class used as an alternate loading path for VR,
back before any of the menus worked in VR mode.  It skipped
most of the regular menu flow and dumped the player directly
into a game.  Since the UI works in VR mode now, this is no
longer needed, and can be removed.

Change-Id: If8fcd9e21a16e46c1b33810bc0fca37bf3a39cc6