Fixed some camera issues.
Camera now starts focused on the start point when loading
a level in the editor.  This should prevent the situation
where a level is built far away from the origin point.
(leaving an empty space screen when the level is loaded.)

Also fixed the bug where the keyboard camera controls
only worked when the camera-drag tool was not selected.

Change-Id: I8ad679f1d0ecf55263195629225618b1fe83a213