Increased version number to 22.
Encountered a bug in the tap presubmit tests, where they are
incorrectly checking the APK version field, instead of the
MinSdkVersion field.  (Which was causing the presubmit checks
to fail for MechaHamster.)  The TAP Bug is still being fixed,
but this change is a valid workaround in the mean time.

b/37430791

Change-Id: I21fd67d77a7c22a244208406aa580f9791b24f16