Added the ability to save and play back replays.
You can now save replays of gameplay.  The option is disabled
by default, but the code is there.  Also included a gameplay
controller that allows replays to be played back.

Also ended up needing to convert some of the logic in gameplay
to fire on fixed updates instead of the regular floating-timestep
update, in order to make the replays more deterministic with
the physics system.

Change-Id: I365e1f50f74b4b2256243f8e6c034d759ba9c639