Added a TEST_LOOP intent to android build.
Set up an intent to the android build, TEST_LOOP.  When
invoked on startup, it will sign in anonymously, (or with
the current default signin if one exists) load a level, and
play through it using replay data.  After completion of the
level, the app exits.

This does not affect non-android gameplay, nor app launches
that are not invoked through the TEST_LOOP intent.

Tested on a Pixel phone.

Change-Id: Ic406d3c4c33cb98a38dc717f9f5e68ecc1ed40c2