Added updated license dialog for launch.
Added the updated license file for launching as a standalone
app.  (It's a longer, more extensive list than the one used
for open source launch.)

Unfortunately, this goes over unity's limit of the maximum number
of characters per text field, so this change also includes a
text field with custom behavior, which breaks input text into
multiple text fields for the scroll view.

Change-Id: I9bbae74233cf8791e9aa75944f2fea082a49f130