Made loading user data more robust.
Game now has a distinction between "no data created yet" vs.
"can't retrieve data."  If the user is signed in (according
to auth) but data can't be retrieved from the database, the
game now uses the same "not signed in" mode that happens if
they refuse from the sign-in screen.

The game then rechecks every time they return to the main
menu, to see if it can get user data yet.  (And if so, restores
full functionality.)

As a result of this change, WaitForDBLoad() now has slightly
different behavior - previously it would return "success = false"
if no data was returned.  (because the request pointed to a
nonexistent record.)  Now, success is returned as true as long
as the request completes successfully.  Consumers of the data
can no longer assume that a successful request implies that
request.data is non-null.  (And need to add a null check.)

Change-Id: Ib8091d91bd6f2bcbaaa9a2a3689ad5892fb01e06