Added a "back" button during gameplay.
There is now a small button in the upper right corner
of the screen during gameplay and edit modes.  This
should provide a more convenient way to exit beyond the
android "back" button.  (Which is normally hidden anyway.)

Also removed the ios-only back button.  (Which was implemented
using the old UI system.)

Change-Id: I21a5d48c7f90a98f5011df7f479976231da89708