Change AudioOnCollision to toggle the AudioSource.
Looking at the profiler, the AudioManager update seems to be taking
longer than it should when no sounds are playing. To improve this,
this changes the AudioOnCollision script to enable/disable the
AudioSource to only be active when the clip is playing. Note, this
only works because every tile that uses AudioOnCollision does not try
to play audio any other way.

Change-Id: Idef6ae9fda2406da686bb17ef5bdc19878309771