Added custom HTML invite message.
Switched to HTML-based email invites.  This gives them a better
message, and the ability to view them in the editor.

Also realized that I was misusing the TitleText field for
invites, and fixed it.

Added the new EmailContentHTML and EmailSubjectText constants to
the string constants file, and set them up for remote config.

Change-Id: I9c6f161bd0261c2b32706915814adbbf4725d1b3