Cloud function to cleanup database and the storage
Add a cloud function to be triggered whenever a new score record is
added to the database.  This function limits the number of records
stored in the database as well as remove unreference replay data from
the storage.

Move the all the rank records to Leaderboard/Map/{MapType}/{MapID}/Top/
so that the rank data won't be mixed with the map information like
BonusMap.  Unfortunately we still need to keep 'Time' in BonusMap and
MapList due to legacy.

Update database rule: add access to Leaderboard and allow uploading time
in current version for BonusMap

Change-Id: Iff663e4f2e9f0000f562296546be16669662020b