Falling off the world doesn't destroy the world anymore. This doesn't make sense for multiplayer. Instead, that player is simply reset back to one of the spawn spots.
StartPosition is not the boss of Player, so should not be destroying the player when the map Reset() is called.
Unet's NetworkStartPosition is the only thing that matters for the network spawn position. Everything else is for single player.
- Thus, the prefab was changed to have NetworkStartPosition be required for StartPosition component.
The CustomNetworkManager needs to know about the player object in order to teleport it to the start position if the player dies.

Includes stubs of transitionary code to multiple players.