Add Firebase Crashlytics to Mechahamster.
Add the Firebase Crashlytics SDK to Mechahamster to start
reporting crashes to the Mechahamster dashboard. In addition
add functionality to the app to force a crash using a new
"Self Destruct" button in the Settings Menu. When this button
is clicked 3 times, the user will log one of 5 exception types
as non-fatal to Crashlytics and see an error menu that fades
to nothing.

Change-Id: I2689f42911bb5f7f592d0150334f4485ca727dc6