Incorporate fix to LeaderboardController.
The LeaderboardController's "GetInitialTopScores" method was hanging in some Android apps when checking if the query result "ChildrenCount" was 0 (likely an SDK issue).
The fix changed that to "HasChildren," which does not hang.
Also in the change, when creating a UserScore directly from a DataSnapshot record, adds some protection around the record having the required fields, and quitting early if not.

Change-Id: Ia111a9015715d8f57942ea1484ebba2683257f13