If there is no gyroscope, fall back to keyboard input
This is an imperfect attempt to better support ChromeOS. Some devices will not have a gyroscope but will have a keyboard for input. A full fix would probably also require touchscreen input (for phones without a gyro) and some sort of user toggle between gyro, touchscreen, and keyboard (for instance: a large chromebook would be very unweidly to tilt to play even if it has a gyroscope).